requirements.txt
# Crypto Bot v3.0 - PyQt6 Version
# Compatible macOS 10.13+

# API & Requests
requests>=2.31.0
urllib3<2.0  # Fix LibreSSL compatibility

# Data Processing
pandas>=2.0.0
numpy>=1.24.0

# Visualization
matplotlib>=3.7.0
plotly>=5.14.0

# GUI - PyQt6 instead of customtkinter
PyQt6>=6.4.0
PyQt6-Charts>=6.4.0

# Database
sqlalchemy>=2.0.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Utilities
pytz>=2023.3
python-dateutil>=2.8.2

# Optional
pytest>=7.4.0
black>=23.0.0
